feat(application): add service for creating cars

+++ src/main/java/ru/lebedev/dealership/domain/car/valueobjects/EngineCapacity.java
@@ -1,6 +1,6 @@
 package ru.lebedev.dealership.domain.car.valueobjects;
 
-public record EngineCapacity(long liters) {
+public record EngineCapacity(double liters) {
     public EngineCapacity {
         if (liters <= 0) {
             throw new IllegalArgumentException("Capacity must be positive");
